% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crema_model.R
\name{ATAC_weighted_RNA_matrix}
\alias{ATAC_weighted_RNA_matrix}
\title{ATAC weighted RNA matrix for the TFs}
\usage{
ATAC_weighted_RNA_matrix(
  target,
  TFs,
  regions_str,
  motifs,
  exp_mtx,
  fragment_object,
  cells,
  genome = BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10,
  motif_match_p_cutoff = 5e-05,
  site_extension = 200,
  smooth = F,
  smooth_cutoff = 1,
  neighbor_mtx = NULL
)
}
\arguments{
\item{neighbor_mtx}{}
}
\value{

}
\description{
ATAC weighted RNA matrix for the TFs
}
