% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crema_model.R
\name{ATAC_weighted_tf_model_highres}
\alias{ATAC_weighted_tf_model_highres}
\title{CREMA model for selected gene}
\usage{
ATAC_weighted_tf_model_highres(
  target,
  TFs,
  regions_str,
  motifs,
  exp_mtx,
  fragment_object,
  cells = NULL,
  genome = BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10,
  motif_match_p_cutoff = 5e-05,
  site_extension = 200,
  return_val = c("list", "vector", "df"),
  regression_method = c("ols", "poisson"),
  smooth = F,
  smooth_cutoff = 1,
  neighbor_mtx = NULL,
  cell_grouping = F,
  cell_grouping_clusters = NULL
)
}
\arguments{
\item{motif_match_p_cutoff}{the p value cutoff for matching motifs, i.e. the \code{p.cutoff} argument for the \code{motifmatchr::matchMotifs} function.}

\item{site_extension}{the extension window surrounding the motif site for calculating chromatin accessibility (ATAC Tn5 insert sites)}

\item{return_val}{}
}
\description{
Run the crema model on one gene
}
