% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_exp_mtx_genes}
\alias{filter_exp_mtx_genes}
\title{Title}
\usage{
filter_exp_mtx_genes(
  exp_mtx,
  proportion_cells_detected = 0.005,
  min_cell_number = 3,
  total_umi_cutoff = "default",
  gene_names_remove_pattern = "^MT-"
)
}
\arguments{
\item{gene_names_remove_pattern}{}
}
\description{
Title
}
