% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_selection.R
\name{filter_motifs_by_tf_names}
\alias{filter_motifs_by_tf_names}
\title{filter motifs by the selected TFs}
\usage{
filter_motifs_by_tf_names(motifs, tfs = NULL)
}
\arguments{
\item{tfs}{}
}
\value{

}
\description{
motifs names are stored as "TF_motifID"
}
