% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crema_model.R
\name{get_motif_access_mtx}
\alias{get_motif_access_mtx}
\title{Get access matrix for each motif}
\usage{
get_motif_access_mtx(
  regions,
  motifs,
  fragment_object,
  tfs = NULL,
  cells = NULL,
  site_extension = 200,
  genome = BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10,
  p.cutoff = 5e-05,
  bg = "even",
  region_extension = 0
)
}
\arguments{
\item{fragment_object}{The fragment object of a \code{Seurat}/\code{Signac} object.}

\item{region_extension}{}
}
\value{

}
\description{
For each TF, get the ATAC accessibility matrix of each motif loccations of the TF
}
