% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crema_model.R
\name{regression_model}
\alias{regression_model}
\title{Title}
\usage{
regression_model(
  x,
  y,
  cell_grouping = F,
  cell_grouping_clusters = NULL,
  regression_method = c("ols", "poisson")
)
}
\arguments{
\item{regression_method}{}
}
\value{

}
\description{
Title
}
