% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_selection.R
\name{scan_TF_positions}
\alias{scan_TF_positions}
\title{Find the matching positions of the motifs}
\usage{
scan_TF_positions(
  regions,
  motifs,
  genome = BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10,
  p.cutoff = 5e-05,
  bg = "even",
  region_extension = 0,
  reduce = F,
  reduce_gap = 0
)
}
\arguments{
\item{regions}{a GRanges object of regions to search for motifs. Will be used as the \code{subject} argument for the \code{matchMotifs} function}

\item{motifs}{a list of position weight matrices. Will be use as the \code{pwm} argument for the \code{matchMotifs} function}

\item{genome}{a BSgenome object. Will be used as the \code{genome} argument for the \code{matchMotifs} function}

\item{p.cutoff}{the \code{p.cutoff} argument for the \code{matchMotifs} function}

\item{bg}{the \code{bg} argument for the \code{matchMotifs} function}

\item{region_extension}{whether to extend the region to search for motifs.}

\item{reduce}{(Not implemented yet) whether to combine / reduce binding sites for the same TF if they are close to each other}

\item{reduce_gap}{the gap allowed for reducing the binding sites.}
}
\value{
a Granges object of the positions of all the mapped motif (motif with at least one match). The GRanges object has metadata column for the TF names
}
\description{
A wrapper for the \code{motifmatchr::matchMotifs} function to find motif match (potential binding sites) locations in a given region
}
