% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_selection.R
\name{select_proximal_regions}
\alias{select_proximal_regions}
\title{Select candidate region for each gene}
\usage{
select_proximal_regions(
  genes,
  gene_body_gr,
  window_up = 1e+05,
  window_down = window_up,
  manual_truncate = F,
  genome = NULL,
  gene_region_inclusion = c("all", "all_noncoding", "closest_noncoding"),
  self_region_inclusion = T
)
}
\arguments{
\item{genes}{list of genes to run CREMA on}

\item{gene_body_gr}{a GRanges object of the genomic positions of each gene. The gene names are in the "gene_name" metadata}

\item{window_up}{search window in the upstream of the TSS}

\item{window_down}{search window in the downstream of the TSS}

\item{manual_truncate}{default FALSE. if the supplied gene_body_gr doesn't have seqinfo, need to manually truncate the regions so they don't surpass the boundaries of chromosomes}

\item{genome}{is manual_truncate = TRUE, supply the genome to get the length of each chromosome}

\item{gene_region_inclusion}{see details}

\item{self_region_inclusion}{default TRUE. Where to include the coding region of the gene itself.}
}
\value{
a list of GRanges objects named by the gene names. Each GRanges object correspond to the selected candidate region of the gene.
}
\description{
This function selects the cnadidate region for each gene to run CREMA model on. Deefault is +/- 100kb around the TSS
}
